requirements.txt
Jinja2
PyYAML
python-gilt
requests
ruamel.yaml
